easing (written but needs to be fixed)

+++ cocos2d/external/tinyxml2/tinyxml2.cpp
@@ -610,7 +610,7 @@ void XMLNode::SetValue( const char* str, bool staticMem )
     }
 }
 
-void XMLNode::SetText(int v)
+void XMLNode::SetText(float v)
 {
     char buf[200];
     XMLUtil::ToStr(v, buf, 200);
